Beginning work on creating a custom namedlist to handle circular references
Have a working solution for repr and string printing. Need to clean up the syntax that it is printed out but I want to see what it is like for JSON encoder first and how to do it. Also, want to look at standard for how class should be printed out