Had to copy/paste JSONEncoder to make changes for circular references
I'm not a fan of having to take source code and copy it over and alter it. This is very confusing but in this case I think it's required for namedlist and JSONEncoder to appropiately handle circular references. Also, I doubt these classes get much changes in them so I won't have to worry too much about keeping them up to date.