Remove old README and add new RST README
Switched to RST because that is primarily what PyPi supports for the long description. The setup.py file uses the README.rst as the long description. That way whether a user gets here from PyPi or GitHub, they get the same information and there will be no discrepencies.

To be clear, PyPi recently started supporting MarkDown in March 2018, but that is too recent for me. I would rather use the tried and true method for awhile in case there are bugs or issues.