Fix issues with setuptools and finding pypdn folder modules
Was incorrectly using py_modules to specify the module but that is only meant for single files to be included. Typically, the best way is to use find_packages function for the packages and let that automatically find the files to use.

Also, include tests and the tests data in install in case one wants to perform the unit tests. Also gives user access to the data in case they want to experiment.

Fix issue with setuptools and finding the pypdn folder modules